Create module for form_controls
